category/games/2026/02/20260210-223343-vector-balloon-fight-fish-hazard/pyproject.toml
[project]
name = "vector-balloon-fight-fish-hazard"
version = "0.1.0"
description = "Master buoyancy and avoid the giant fish in this gravity-defying aerial duel."
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
